a=int(input()) area=3.14*a*a print(f"{area:.2f}")

+++ Area_of_a_circle.py
@@ -1,3 +1,3 @@
-a=float(input())
+a=int(input())
 area=3.14*a*a
 print(f"{area:.2f}")
